ML Model/ml_model/requirements.txt
Flask==2.2.2
google-generativeai==0.0.7
